+++ R/soft-rule.R
@@ -37,11 +37,18 @@ eps_min <- suffix("_eps_min")
 #'
 #' @param values named list of values.
 #' @param weights named numeric of equal length as values.
+#' @param delta_names alternative names for binary variables. (used for log)
 #' @param ... not used
-expect_values <- function(values, weights, ...){
+#' @keywords internal
+expect_values <- function(values, weights, delta_names = NULL, ...){
   if (missing(weights)){
     weights <- rep(1, length(values))
     names(weights) <- names(values)
+  }
+
+  if (is.null(delta_names)){
+    delta_names <- names(values)
+    names(delta_names) <- delta_names
   }
 
   stopifnot(
@@ -61,7 +68,10 @@ expect_values <- function(values, weights, ...){
     b <- lin_values[[n]]
     w <- weights[n]
     if (is.finite(w)){
-      soft_lin_rule(mip_rule(a, op="<=", b =  b, rule = n, weight = w))
+      n_ub <- paste0(n, "_ub")
+      soft_lin_rule( mip_rule(a, op="<=", b =  b, rule = n_ub, weight = w)
+                   , name = delta_names[n]
+                   )
     } else {
       mip_rule(a, op = "==", b = b, rule = n, weight = Inf)
     }
@@ -75,7 +85,10 @@ expect_values <- function(values, weights, ...){
     b <- lin_values[[n]]
     w <- weights[n]
     if (is.finite(w)){
-      soft_lin_rule(mip_rule(-a, op = "<=", b = -b, rule = n, weight = w))
+      n_lb <- paste0(n, "_lb")
+      soft_lin_rule( mip_rule(-a, op = "<=", b = -b, rule = n_lb, weight = w)
+                   , name = delta_names[n]
+                   )
     } else {
       NULL
     }
@@ -88,10 +101,16 @@ expect_values <- function(values, weights, ...){
     b <- 1
 
     if (is.logical(value)){
+      # rewrite (only need one column)
       names(a) <- n
-      if (!value){
+      if (!isTRUE(value)){
         a <- -a
         b <- 0
+      }
+      # if NA, just skip this constraint, should this also hold for categories?
+      if (is.na(value)){
+        a[] <- 0 # effectively set delta to 1
+        b <- 1
       }
     }
     soft_cat_rule(mip_rule(a, op = "==", b = b, rule = n, weight = weights[n], type=sapply(a, function(x) "binary")))