+++ tests/testthat/test-soft-rule.R
@@ -46,14 +46,14 @@ describe("soft_lin_rule",{
     sr <- expect_values(values)[[1]]
     expect_equal(sr$a, c(x=1, .delta_x = -1e7))
     expect_equal(sr$op, "<=")
-    expect_equal(sr$rule, "x")
+    expect_equal(sr$rule, "x_ub")
     expect_equal(sr$b, 42)
     expect_equal(sr$type, c(x="double", .delta_x = "binary"))
 
     sr <- expect_values(values)[[2]]
     expect_equal(sr$a, c(x=-1, .delta_x = -1e7))
     expect_equal(sr$op, "<=")
-    expect_equal(sr$rule, "x")
+    expect_equal(sr$rule, "x_lb")
     expect_equal(sr$b, -42)
   })
 
@@ -67,4 +67,8 @@ describe("soft_lin_rule",{
     expect_equal(sr$rule, "A")
   })
 
+  it ("handles NA logical values correctly",{
+    values <- list(L = NA)
+    expect_values(values)
+  })
 })